refactor start-app
