+++ packages/yoshi/src/commands/start-app.js
@@ -19,167 +19,11 @@ if (cliArgs.production) {
   process.env.NODE_ENV = 'production';
 }
 
-const fs = require('fs-extra');
-const chalk = require('chalk');
-const openBrowser = require('./utils/open-browser');
-const project = require('yoshi-config');
-const { PORT } = require('../constants');
-const {
-  createClientWebpackConfig,
-  createServerWebpackConfig,
-} = require('../../config/webpack.config');
-const {
-  createCompiler,
-  createDevServer,
-  waitForCompilation,
-} = require('../webpack-utils');
-const ServerProcess = require('../server-process');
-const detect = require('detect-port');
-const { watchPublicFolder } = require('./utils/copy-assets');
 const rootApp = require('yoshi-config/root-app');
+const startSingleApp = require('./utils/start-single-app');
 
-const host = '0.0.0.0';
-
-const https = cliArgs.https || project.servers.cdn.ssl;
-
-module.exports = async ({ app = rootApp }) => {
-  // Clean tmp folders
-  await Promise.all([fs.emptyDir(app.BUILD_DIR), fs.emptyDir(app.TARGET_DIR)]);
-
-  // Copy public to statics dir
-  if (await fs.pathExists(app.PUBLIC_DIR)) {
-    // all files in `PUBLIC_DIR` are copied initially as Chokidar's `ignoreInitial`
-    // option is set to false
-    watchPublicFolder();
-  }
-
-  // Generate an available port for server HMR
-  const hmrPort = await detect();
-
-  const clientConfig = createClientWebpackConfig({
-    app,
-    isDebug: true,
-    isAnalyze: false,
-    isHmr: project.hmr,
-  });
-
-  const serverConfig = createServerWebpackConfig({
-    app,
-    isDebug: true,
-    isHmr: true,
-    hmrPort,
-  });
-
-  // Configure compilation
-  const multiCompiler = createCompiler([clientConfig, serverConfig], { https });
-  const compilationPromise = waitForCompilation(multiCompiler);
-
-  const [clientCompiler, serverCompiler] = multiCompiler.compilers;
-
-  // Start up server process
-  const serverProcess = new ServerProcess({
-    serverFilePath: cliArgs.server,
-    hmrPort,
-    app,
-  });
-
-  // Start up webpack dev server
-  const devServer = await createDevServer(clientCompiler, {
-    publicPath: clientConfig.output.publicPath,
-    https,
-    host,
-    app,
-  });
-
-  serverCompiler.watch({ 'info-verbosity': 'none' }, async (error, stats) => {
-    // We save the result of this build to webpack-dev-server's internal state so the last
-    // server build results are sent to the browser on every refresh
-    //
-    // https://github.com/webpack/webpack-dev-server/blob/master/lib/Server.js#L144
-    devServer._stats = stats;
-
-    const jsonStats = stats.toJson();
-
-    // If the spawned server process has died, restart it
-    if (serverProcess.child && serverProcess.child.exitCode !== null) {
-      await serverProcess.restart();
-
-      // Send the browser an instruction to refresh
-      await devServer.send('hash', jsonStats.hash);
-      await devServer.send('ok');
-    }
-    // If it's alive, send it a message to trigger HMR
-    else {
-      // If there are no errors and the server can be refreshed
-      // then send it a signal and wait for a responsne
-      if (serverProcess.child && !error && !stats.hasErrors()) {
-        const { success } = await serverProcess.send({});
-
-        // HMR wasn't successful, restart the server process
-        if (!success) {
-          await serverProcess.restart();
-        }
-
-        // Send the browser an instruction to refresh
-        await devServer.send('hash', jsonStats.hash);
-        await devServer.send('ok');
-      } else {
-        // If there are errors, show them on the browser
-        if (jsonStats.errors.length > 0) {
-          await devServer.send('errors', jsonStats.errors);
-        } else if (jsonStats.warnings.length > 0) {
-          await devServer.send('warnings', jsonStats.warnings);
-        }
-      }
-    }
-  });
-
-  console.log(chalk.cyan('Starting development environment...\n'));
-
-  // Start up webpack dev server
-  await new Promise((resolve, reject) => {
-    devServer.listen(project.servers.cdn.port, host, err =>
-      err ? reject(err) : resolve(devServer),
-    );
-  });
-
-  // Wait for both compilations to finish
-  try {
-    await compilationPromise;
-  } catch (error) {
-    // We already log compilation errors in a compiler hook
-    // If there's an error, just exit(1)
-    process.exit(1);
-  }
-
-  ['SIGINT', 'SIGTERM'].forEach(sig => {
-    process.on(sig, () => {
-      serverProcess.end();
-      devServer.close();
-      process.exit();
-    });
-  });
-
-  try {
-    await serverProcess.initialize();
-  } catch (error) {
-    console.log();
-    console.log(
-      chalk.red(`Couldn't find a server running on port ${chalk.bold(PORT)}`),
-    );
-    console.log(
-      chalk.red(
-        `Please check that ${chalk.bold(
-          cliArgs.server,
-        )} starts up correctly and that it listens on the expected port`,
-      ),
-    );
-    console.log();
-    console.log(chalk.red('Aborting'));
-    process.exit(1);
-  }
-
-  openBrowser(cliArgs.url || project.startUrl || `http://localhost:${PORT}`);
+module.exports = async () => {
+  await startSingleApp(rootApp, cliArgs);
 
   return {
     persistent: true,
